backend/requirements.txt
absl-py==2.4.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@ded190a052fdf4585bd685cee5bc96e0310d2c93
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cycler==0.12.1
filelock==3.25.0
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
ftfy==6.3.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
outcome==1.3.0.post0
packaging==26.0
pillow==12.1.1
protobuf==7.34.0
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
regex==2026.2.28
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
selenium==4.41.0
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
wsproto==1.3.2
